Move config structs into config.go.
Part of https://github.com/letsencrypt/boulder/issues/1052. I'll be adding some
new config structs, and want everything in a consistent place.
